Some structural refactoring
