+++ schema-registry/authorizer/authorizer-agent/src/test/resources/schema-registry-policies-tag.json
