+++ schema-registry/authorizer/pom.xml
@@ -3,21 +3,19 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>registries</artifactId>
+        <artifactId>schema-registry</artifactId>
         <groupId>com.hortonworks.registries</groupId>
         <version>0.8.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>schema-registry</artifactId>
+    <artifactId>authorizer</artifactId>
     <packaging>pom</packaging>
     <modules>
-        <module>common</module>
-        <module>client</module>
-        <module>core</module>
-        <module>serdes</module>
-        <module>rest-service</module>
-        <module>authorizer</module>
+        <module>authorizer-core</module>
+        <module>authorizer-agent</module>
+        <module>ranger-plugin</module>
     </modules>
+
 
 </project>
