+++ schema-registry/authorizer/ranger-plugin/ranger-authorizer/conf/ranger-policymgr-ssl.xml
