+++ schema-registry/authorizer/ranger-plugin/ranger-authorizer/src/main/assembly/binary.xml
