+++ schema-registry/authorizer/ranger-plugin/ranger-authorizer/src/test/java/com/hortonworks/registries/ranger/authorization/schemaregistry/authorizer/util/SecurityContextForTesting.java
