+++ schema-registry/authorizer/ranger-plugin/ranger-authorizer/src/test/resources/ranger-schema-registry-security.xml
