+++ schema-registry/authorizer/ranger-plugin/ranger-authorizer/src/test/resources/schema-registry-policies-tag.json
