+++ schema-registry/rest-service/pom.xml
@@ -187,7 +187,7 @@
       </dependency>
         <dependency>
             <groupId>com.hortonworks.registries</groupId>
-            <artifactId>registry-authorization-agent</artifactId>
+            <artifactId>authorizer-agent</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
